models API/pestsApi/requirements.txt
fastapi
uvicorn
ultralytics
torch
torchvision
opencv-python
pillow
numpy
python-multipart